chore: move networking code to the YAT.Net namespace
