+++ addons/yat/src/Net/EPingStatus.cs
@@ -1,0 +1,8 @@
+namespace YAT.Net;
+
+public enum EPingStatus : ushort
+{
+    Success = 0,
+    Unsupported = 1,
+    Unknown = 2
+}
